Separated Http specific client and 'Simple' client
