+++ channels/tests/__init__.py
@@ -1,1 +1,2 @@
 from .base import ChannelTestCase, Client, apply_routes  # NOQA isort:skip
+from .http import HttpClient
